Add os-vif plugin
Note: this requires unix-socket based mm-ctl api, which
is only available for MidoNet 5.2.1 and later. [1]
It means, once nova enabled the support [2] for this,
it will be incompatible with older versions of MidoNet agents.
If we want to support those versions of MidoNet, we can make
this execute the mm-ctl command instead of talking to the
socket directly.

[1] https://github.com/midonet/midonet/blob/master/docs/mm-ctl.md
[2] I4a1d8f5b2f100f86dd1ee7dca3983881fb86e143

Closes-Bug: #1544831
Change-Id: I6a3d52b606e43a9d00ec1a7c7db86ee32c1511ca